Exceptions are now used in logic

+++ OGAOE7_HFT_2021221.Logic/Logic.cs
@@ -1,4 +1,5 @@
-﻿using OGAOE7_HFT_2021221.Repository;
+﻿using OGAOE7_HFT_2021221.Logic.Exceptions;
+using OGAOE7_HFT_2021221.Repository;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -15,13 +16,15 @@ namespace OGAOE7_HFT_2021221.Logic
             this.repo = repo;
         }
 
-        public void Create(T newItem)
+        public virtual void Create(T newItem)
         {
             repo.Create(newItem);
         }
 
         public IEnumerable<T> Read(int id)
         {
+            if (id < 0) throw new UnsupportedValueException(id);
+
             return new List<T> { repo.Read(id) };
         }
 
@@ -32,12 +35,27 @@ namespace OGAOE7_HFT_2021221.Logic
 
         public void Update(T newItem)
         {
-            repo.Update(newItem);
+            try
+            {
+                repo.Update(newItem);
+            }
+            catch
+            {
+                throw new OperationFailedException<T>((newItem) => repo.Update(newItem));
+            }
         }
 
         public void Delete(int id)
         {
-            repo.Delete(id);
+            if (id < 0) throw new UnsupportedValueException(id);
+            try
+            {
+                repo.Delete(id);
+            }
+            catch
+            {
+                throw new OperationFailedException<int>((id) => repo.Delete(id));
+            }
         }
 
         public abstract IEnumerable<string> MainData(int id);
